pyproject.toml
[tool.poetry]
name = "iq-eq-testapp"
version = "0.1.0"
description = ""
authors = ["Almukhan Chingiz <chingizpain@gmail.com>"]
readme = "README.md"
packages = [{include = "iq_eq_testapp"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
